+++ src/js/controller.js
@@ -5,16 +5,6 @@ import "regenerator-runtime/runtime";
 
 //const recipeContainer = document.querySelector(".recipe");
 
-const timeout = function (s) {
-    return new Promise(function (_, reject) {
-        setTimeout(function () {
-            reject(
-                new Error(`Request took too long! Timeout after ${s} second`)
-            );
-        }, s * 1000);
-    });
-};
-
 // https://forkify-api.herokuapp.com/v2
 
 ///////////////////////////////////////
@@ -24,7 +14,6 @@ const controlRecipes = async function () {
         // * LOADING RECIPE
         // getting id of hash
         const id = window.location.hash.slice(1);
-        console.log(id);
         if (!id) return;
 
         recipeView.renderSpinner();
