+++ src/js/model.js
@@ -1,4 +1,6 @@
 import { async } from "regenerator-runtime";
+import { API_URL } from "./config.js";
+import { getJSON } from "./helpers.js";
 
 export const state = {
     recipe: {},
@@ -6,14 +8,8 @@ export const state = {
 
 export const loadRecipe = async function (id) {
     try {
-        const res = await fetch(
-            `https://forkify-api.herokuapp.com/api/v2/recipes/${id}`
-        );
-        const data = await res.json();
-        // Error message
-        if (!res.ok)
-            throw new Error(`${data.message} (status : ${res.status})`);
-        // recipe new obj
+        const data = await getJSON(`${API_URL}/${id}`);
+
         const { recipe } = data.data;
         state.recipe = {
             id: recipe.id,
@@ -26,6 +22,6 @@ export const loadRecipe = async function (id) {
             ingredients: recipe.ingredients,
         };
     } catch (error) {
-        console.error(error);
+        console.error(`SOMTHING WENT WRONG! ${error}`);
     }
 };
